+++ src/main/kotlin/org/appjam/smashing/global/auth/jwt/components/JwtProvider.kt
@@ -1,7 +1,10 @@
-package org.appjam.smashing.global.auth.jwt
+package org.appjam.smashing.global.auth.jwt.components
 
-import org.appjam.smashing.global.auth.jwt.JwtGenerator.Companion.ROLES_KEY
-import org.appjam.smashing.global.auth.jwt.JwtGenerator.Companion.TYPE_KEY
+import org.appjam.smashing.global.auth.jwt.components.JwtGenerator.Companion.ROLES_KEY
+import org.appjam.smashing.global.auth.jwt.components.JwtGenerator.Companion.TYPE_KEY
+import org.appjam.smashing.global.auth.jwt.dto.TokenDto
+import org.appjam.smashing.global.auth.jwt.enums.TokenType
+import org.appjam.smashing.global.auth.security.CustomUserDetails
 import org.appjam.smashing.global.exception.CustomException
 import org.appjam.smashing.global.exception.ErrorCode
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
