+++ src/main/kotlin/org/appjam/smashing/global/auth/jwt/config/JwtProperties.kt
@@ -1,4 +1,4 @@
-package org.appjam.smashing.global.auth.jwt
+package org.appjam.smashing.global.auth.jwt.config
 
 import org.springframework.boot.context.properties.ConfigurationProperties
