+++ src/main/kotlin/org/appjam/smashing/global/auth/jwt/enums/TokenType.kt
@@ -1,4 +1,4 @@
-package org.appjam.smashing.global.auth.jwt
+package org.appjam.smashing.global.auth.jwt.enums
 
 enum class TokenType {
     ACCESS_TOKEN, REFRESH_TOKEN
