+++ src/main/kotlin/org/appjam/smashing/global/auth/jwt/handler/JwtAuthenticationEntryPoint.kt
@@ -1,27 +1,33 @@
-package org.appjam.smashing.global.auth.exception
+package org.appjam.smashing.global.auth.jwt.handler
 
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.appjam.smashing.global.exception.CustomException
 import org.appjam.smashing.global.exception.ErrorCode
 import org.springframework.beans.factory.annotation.Qualifier
-import org.springframework.security.access.AccessDeniedException
-import org.springframework.security.web.access.AccessDeniedHandler
+import org.springframework.security.core.AuthenticationException
+import org.springframework.security.web.AuthenticationEntryPoint
 import org.springframework.stereotype.Component
 import org.springframework.web.servlet.HandlerExceptionResolver
 
 @Component
-class JwtAccessDeniedHandler(
+class JwtAuthenticationEntryPoint(
     @Qualifier("handlerExceptionResolver")
     private val resolver: HandlerExceptionResolver,
-) : AccessDeniedHandler {
+) : AuthenticationEntryPoint {
 
-    override fun handle(
+    override fun commence(
         request: HttpServletRequest,
         response: HttpServletResponse,
-        accessDeniedException: AccessDeniedException,
+        authException: AuthenticationException,
     ) {
-        val exception = CustomException(ErrorCode.FORBIDDEN)
+        val exception = request.getAttribute(EXCEPTION_KEY) as? CustomException ?: CustomException(ErrorCode.UNAUTHORIZED)
+
         resolver.resolveException(request, response, null, exception)
     }
+
+    companion object {
+        const val EXCEPTION_KEY = "exception"
+    }
+
 }
